Migrate Cron Jobs to new resource tracking parser
