+++ parser/testdata/cron_job_definition_init.txt
@@ -1,6 +1,6 @@
 # Verify the cron job definition
 ! parse
-stderr 'cron job must be defined as a package global variable'
+stderr 'A cron job can not be declared here, they can only be declared in a package level variable. For more information please see https://encore.dev/docs/develop/cron-jobs'
 
 -- svc/svc.go --
 package svc
